Desenvolvido o incremento da operação através de TDD.

+++ src/main/java/one/digitalinnovation/beerstock/controller/BeerController.java
@@ -1,6 +1,7 @@
 package one.digitalinnovation.beerstock.controller;
 
 import lombok.AllArgsConstructor;
+import one.digitalinnovation.beerstock.controller.*;
 import one.digitalinnovation.beerstock.dto.BeerDTO;
 import one.digitalinnovation.beerstock.dto.QuantityDTO;
 import one.digitalinnovation.beerstock.exception.BeerAlreadyRegisteredException;
@@ -9,6 +10,7 @@ import one.digitalinnovation.beerstock.exception.BeerStockExceededException;
 import one.digitalinnovation.beerstock.service.BeerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -18,6 +20,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+
+
 
 import javax.validation.Valid;
 import java.util.List;
@@ -57,6 +61,6 @@ public class BeerController implements BeerControllerDocs {
 
     @PatchMapping("/{id}/increment")
     public BeerDTO increment(@PathVariable Long id, @RequestBody @Valid QuantityDTO quantityDTO) throws BeerNotFoundException, BeerStockExceededException {
-        return beerService.increment(id, (Integer) quantityDTO.getQuantity());
+        return beerService.increment(id, quantityDTO.getQuantity());
     }
 }
